Refactor project constants
