+++ src/commands/createNewProject/CSharpScriptProjectCreator.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { TemplateFilter } from "../../constants";
 import { localize } from "../../localize";
-import { TemplateFilter } from "../../ProjectSettings";
 import { funcHostTaskId, funcHostTaskLabel } from "./IProjectCreator";
 import { ScriptProjectCreatorBase } from './ScriptProjectCreatorBase';
