+++ src/commands/createNewProject/JavaProjectCreator.ts
@@ -7,8 +7,8 @@ import * as fse from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
 import { InputBoxOptions } from 'vscode';
+import { ProjectRuntime, TemplateFilter } from '../../constants';
 import { localize } from "../../localize";
-import { ProjectRuntime, TemplateFilter } from '../../ProjectSettings';
 import { cpUtils } from '../../utils/cpUtils';
 import * as fsUtil from '../../utils/fs';
 import { validateMavenIdentifier, validatePackageName } from '../../utils/javaNameUtils';
