+++ src/templates/TemplateData.ts
@@ -11,8 +11,8 @@ import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext } from 'vscode-azureextensionui';
 import TelemetryReporter from 'vscode-extension-telemetry';
 import { ScriptProjectCreatorBase } from '../commands/createNewProject/ScriptProjectCreatorBase';
+import { ProjectLanguage, ProjectRuntime, TemplateFilter } from '../constants';
 import { localize } from '../localize';
-import { ProjectLanguage, ProjectRuntime, TemplateFilter } from '../ProjectSettings';
 import { Config } from './Config';
 import { ConfigBinding } from './ConfigBinding';
 import { ConfigSetting } from './ConfigSetting';
